control-panel/backend/pyproject.toml
# Copyright (c) 2026 Diego Silva Morais <lukewaresoftwarehouse@gmail.com>
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

[project]
name = "clawdevs-panel-backend"
version = "0.1.0"
requires-python = ">=3.12,<4.0"
package-mode = false
dependencies = [
    "fastapi==0.135.1",
    "uvicorn[standard]==0.34.0",
    "sqlmodel==0.0.37",
    "alembic==1.18.4",
    "asyncpg==0.30.0",
    "psycopg[binary]==3.2.4",
    "pgvector==0.2.0",
    "redis[hiredis]==5.3.0",
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "httpx==0.28.1",
    "kubernetes==32.0.1",
    "watchdog==6.0.0",
    "fastapi-pagination==0.12.34",
    "sse-starlette==2.3.6",
    "rq==2.6.0",
    "rq-scheduler==0.14.0",
    "python-multipart==0.0.22",
    "pydantic-settings==2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.0",
    "pytest-asyncio==0.25.3",
    "httpx==0.28.1",
    "pytest-cov==6.1.0",
    "anyio==4.9.0",
    "aiosqlite==0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]